[IMP] *: fixup

+++ addons/hr_expense/models/hr_expense_sheet.py
@@ -468,8 +468,8 @@ class HrExpenseSheet(models.Model):
         if 'state' in vals or 'approval_state' in vals:
             # Avoid user with write access on expense sheet in draft state to bypass the validation process
             if (not self.user_has_groups('hr_expense.group_hr_expense_manager')
-                and self.state == 'draft'
-                and (vals.get('state') != 'submit' or vals.get('approval_state') == 'approve')):
+                and all(self.mapped(lambda e: e.state == 'draft')) and vals.get('state') not in ['submit', None]
+                and vals.get('approval_state') not in ['sumbit', None]):
                 raise UserError(_("You don't have the rights to bypass the validation process of this expense report."))
             elif vals.get('state') == 'approve' or vals.get('approval_state') == 'approve':
                 self._check_can_approve()
@@ -709,13 +709,14 @@ class HrExpenseSheet(models.Model):
     def _do_reverse_moves(self):
         self = self.with_context(clean_context(self.env.context))
         moves = self.account_move_ids
-        draft_moves = moves.filtered(lambda m: m.state == 'draft')
-        non_draft_moves = moves - draft_moves
-        non_draft_moves._reverse_moves(
-            default_values_list=[{'invoice_date': fields.Date.context_today(move), 'ref': False} for move in non_draft_moves],
-            cancel=True
-        )
-        draft_moves.unlink()
+        if moves:
+            draft_moves = moves.filtered(lambda m: m.state == 'draft')
+            non_draft_moves = moves - draft_moves
+            non_draft_moves._reverse_moves(
+                default_values_list=[{'invoice_date': fields.Date.context_today(move), 'ref': False} for move in non_draft_moves],
+                cancel=True
+            )
+            draft_moves.unlink()
 
     def _prepare_bills_vals(self):
         self.ensure_one()
